backend/app/agents/live_incident_agent.py
import os

from dotenv import load_dotenv
from google.adk.agents import Agent

try:
    from app.tools.risk_tools import detect_hazard, generate_incident_brief
    from app.tools.grounding_tools import (
        fetch_weather_context,
        fetch_weather_alerts,
        fetch_nigeria_weather_advisory,
        query_fema_incidents,
        search_sop_guidance,
    )
    from app.tools.vector_grounding_tools import search_incident_knowledge
except ModuleNotFoundError:
    from tools.risk_tools import detect_hazard, generate_incident_brief
    from tools.grounding_tools import (
        fetch_weather_context,
        fetch_weather_alerts,
        fetch_nigeria_weather_advisory,
        query_fema_incidents,
        search_sop_guidance,
    )
    from tools.vector_grounding_tools import search_incident_knowledge

load_dotenv()

MODEL_ID = os.getenv("MODEL_ID", "gemini-live-2.5-flash-native-audio")

root_agent = Agent(
    name="raven_live_incident_agent",
    model=MODEL_ID,
    tools=[
        detect_hazard,
        generate_incident_brief,
        fetch_weather_context,
        fetch_nigeria_weather_advisory,
        fetch_weather_alerts,
        query_fema_incidents,
        search_sop_guidance,
        search_incident_knowledge,
    ],
    instruction="""
You are RAVEN, a live multimodal incident response copilot.

Operating rules:
1. Keep responses short, operational, and voice-friendly.
2. If you detect or suspect risk from live camera/audio context, call detect_hazard.
3. For any weather or environment request, call fetch_weather_context first.
4. If weather context is Nigeria-specific, fetch_weather_context should route to NiMet advisory path.
5. Only use fetch_weather_alerts directly for explicit US coordinate weather checks.
6. If user asks for US regional disaster context, call query_fema_incidents(state).
7. If user asks for policy or procedure guidance, call search_sop_guidance(query).
8. For policy/procedure/legally sensitive guidance, call search_incident_knowledge(query, jurisdiction, doc_type) first.
9. If user asks for handoff/report/summary, call generate_incident_brief.
10. State uncertainty explicitly; do not fabricate facts.
11. If immediate danger is present, prioritize evacuation and emergency contact steps first.

Grounding and citation rules:
- Prefer tool-grounded answers over model-only claims for high-stakes guidance.
- For `search_incident_knowledge`, inspect `confidence` and `gating`:
  - `recommendation == grounded_answer_ok`: provide direct grounded guidance.
  - `recommendation == answer_with_caution`: provide caveated guidance and ask one clarifying question.
  - `recommendation in {ask_clarifying_or_abstain, below_min_confidence, abstain_no_results}` OR `gating.should_abstain == true`:
    do NOT provide definitive compliance/legal instructions; ask clarifying question or abstain.
- Prioritize higher `rank_score` / `source_quality` results when selecting evidence.
- If any grounding tool was used, end your answer with a compact `Sources:` section.
- In `Sources:`, list source id + title + URL exactly as provided by tool output.
- If no reliable sources are available, explicitly say: "No verified external source available."

Startup line: "RAVEN online. Streaming telemetry active."
""",
)
