backend/pyproject.toml
[project]
name = "raven-live-agent-backend"
version = "0.1.0"
description = "RAVEN live multimodal incident response backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets==15.0.1",
    "fastapi==0.128.3",
    "python-dotenv==1.2.1",
    "google-adk==1.24.1",
    "google-genai==1.56.0",
    "uvicorn==0.40.0",
    "httpx==0.28.1",
    "google-cloud-vectorsearch>=0.3.0",
]
